Refactor display to astro component
