+++ src/components/SiteTitle.astro
@@ -3,8 +3,8 @@ import type { Props } from '@astrojs/starlight/props';
 import { Image } from 'astro:assets';
 const { siteTitle, siteTitleHref } = Astro.props;
 
-import dark from '../assets/nemo-logo-rusty-bright-nomargin.svg';
-import light from '../assets/nemo-logo-rusty-nomargin.svg';
+import dark from '@assets/nemo-logo-rusty-bright-nomargin.svg';
+import light from '@assets/nemo-logo-rusty-nomargin.svg';
 ---
 
 <a href={siteTitleHref} class="flex gap-0 content-center no-underline site-title-link">
