--- src/components/format_table.astro
@@ -1,20 +1,0 @@
----
-import {type Format, checkIODirection, type Direction, formats} from '../content/lib/formats.ts';
-import type { HTMLAttributes } from 'astro/types';
-
-interface Props extends HTMLAttributes<'div'> {
-  of: keyof typeof formats;
-  direction: Direction,
-}
-
-const {of, direction } = Astro.props;
-const format: Format = formats[of];
----
-{checkIODirection(format, direction) && (
-  <tr>
-    <td><code>{format.name}</code></td>
-    <td><span set:html={format.description}></span></td>
-  </tr>
-  )
-}
-
