+++ src/content/docs/index.mdx
@@ -5,8 +5,8 @@ template: splash
 hero:
   tagline: A graph rule engine for the semantic web
   image:
-    light: "../../assets/nemo-logo-rusty.svg"
-    dark: "../../assets/nemo-logo-rusty-bright.svg"
+    light: "@assets/nemo-logo-rusty.svg"
+    dark: "@assets/nemo-logo-rusty-bright.svg"
   actions:
     - text: Get Started with nemo
       link: /nemo-doc/intro/welcome/
