+++ src/content/docs/reference/exports.mdx
@@ -1,10 +1,9 @@
 ---
 title: Export Directives
 ---
-import { Tabs, TabItem } from '@astrojs/starlight/components';
-import ParameterLayout from '../../../components/parameter_layout.astro';
-import {formats} from '../../lib/formats.ts';
-import FormatTable from '../../../components/format_table.astro';
+import {formats} from '@lib/formats';
+import DisplayFormats from '@components/display_formats.astro';
+import DisplayParameters from '@components/display_parameters.astro';
 
 Export directives are used to write computation results to files.
 
@@ -26,33 +25,13 @@ The output will be gzipped (guessed from file name). Tuples that contain anythin
 
 Currently the following formats are supported:
 
-<table class="w-full table-auto border-separate">
-  <thead>
-    <tr>
-      <th class="pe-8 text-left w-32 whitespace-nowrap">Format</th>
-      <th class="text-left">Parameters</th>
-    </tr>
-  </thead>
-  <tbody>
-    {Object.entries(formats).map(([key, format]) => (
-        <FormatTable of={key} direction='export' />
-      ))}
-  </tbody>
-</table>
+<DisplayFormats formats={formats} direction="export" />
+
 
 ### Available parameters
 
-<Tabs>
-  {Object.entries(formats).map(([key, format]) => (
-    <TabItem label={format.name}>
-      <table>
-        <tbody>
-          <ParameterLayout of={key} direction='export' />
-        </tbody>
-      </table>
-    </TabItem>
-  ))}
-</Tabs>
+<DisplayParameters formats={formats} direction="export" />
+
 
 When using the [Nemo command-line client](/nemo-doc/guides/cli), some cli options are available to override the export directives in the program, to set the output (base) directory, and to control if existing files should be overwritten.
