+++ tsconfig.json
@@ -1,4 +1,14 @@
 {
   "extends": "astro/tsconfigs/strict",
-  "compilerOptions": {"allowImportingTsExtensions": true},
+  "compilerOptions":{
+    "baseUrl": ".",
+    "paths": {
+      "@assets/*": ["src/assets/*"],
+      "@lib/*": ["src/lib/*"],
+      "@components/*": ["src/components/*"],
+      "@layouts/*": ["src/layouts/*"],
+      "@pages/*": ["src/pages/*"],
+    }
+  },
+  "include": ["src"]
 }
